requirements.txt
python-socketio
fastapi
uvicorn
motor